pyproject.toml
[tool.poetry]
name = "MorphoParse"
version = "1.2.dev0"
description = "Parsing, partitioning, and weighting morphological character matrices for phylogenetic analysis."
authors = ["EJ Huang <ejhuang91016@gmail.com>"]
readme = "README.md"
license = "GPL-3.0-or-later"
packages = [
    { include = "morphoparse", from = "src" }
]
[tool.poetry.dependencies]
python = ">=3.7"
biopython = "*"

[tool.poetry.scripts]
morphoparse = "morphoparse.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
